refactor #94: Home에 있던 드롭다운 별도 컴포넌트로 분리
- UserLocationDropdown 이라는 컴포넌트로 분리했습니다.
- extractKeyName 함수 유틸화
