+++ src/utils/extractKeyName.ts
@@ -1,0 +1,5 @@
+export const extractKeyName = (locationName: string | undefined) => {
+  if (!locationName) return;
+  const keyName = locationName.split(' ').at(-1);
+  return keyName;
+};
